python-sample-vscode-flask-tutorial/requirements.txt
Flask
azure-storage-blob
azure-keyvault-secrets
azure-data-tables
azure-identity